Migrations fix toegepast

+++ database/migrations/2022_03_31_123157_articles.php
@@ -14,6 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('articles', function (Blueprint $table) {
+            if(Schema::hasTable('articles')) return;
             $table->bigIncrements('id');
             $table->unsignedBigInteger('category_id');
             $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
